Update 0.0.7
-budynki nie moga być budowane na drogach (chyba że wymaga tego)
-rework (a raczej finalna wersja) budynków populacji
-nowe tymczasowe menu budowania
-więcej informacji przed postawieniem budynku
-zapisywanie meta progresji w systemie
-artefakty do meta progresji (tymczasowo dropi 1 z każdego mobka)

+++ Assets/Main folder/Scripts/Buildings/BeaconEntity.cs
@@ -78,10 +78,10 @@ public class BeaconEntity : BuildingEntity
             else
             {
                 // Jeœli nie staæ na ca³oœæ, pobierz tyle ile jest
-                int availableCoal = ResourceManager.Instance.GetResourceAmount(ResourceType.Coal);
+                float availableCoal = ResourceManager.Instance.GetResourceAmount(ResourceType.Coal);
                 if (availableCoal > 0)
                 {
-                    var partialCost = new Dictionary<ResourceType, int>
+                    var partialCost = new Dictionary<ResourceType, float>
                     {
                         { ResourceType.Coal, availableCoal }
                     };
