+++ Assets/Main folder/Scripts/Save/SaveData.cs
@@ -4,7 +4,7 @@ using System.Collections.Generic;
 [Serializable]
 public class SaveData
 {
-    public int totalArtifacts;
+    public float totalArtifacts;
     // Zapisujemy tylko listê ID odblokowanych ulepszeñ
     public List<string> unlockedUpgradeIDs = new List<string>();
 }
